\name{.ggplot_corInternal}
\alias{.ggplot_corInternal}
\title{Correlation Text...}
\usage{.ggplot_corInternal(xVar, yVar,
    ...)}
\description{Correlation Text
Correlation text printed for a given data set}
\author{Barret Schloerke \email{bigbear@iastate.edu}, Haesung Kim \email{hae0510@iastate.edu}, Di Cook \email{dicook@iastate.edu} and Heike Hofmann \email{hofmann@iastate.edu}}
\keyword{internal}
\arguments{\item{xVar}{x variable}
\item{yVar}{y variable}
\item{...}{arguements to be supplied to geom_text}}
