\name{.plot_types}
\alias{.plot_types}
\title{Plot Types...}
\usage{.plot_types(data)}

\description{
  Plot Types Retrieves the type of plot that should be used for all
  combinations
}
\keyword{internal}
\author{Barret Schloerke \email{bigbear@iastate.edu} and Haesung Kim \email{hae0510@iastate.edu}}
\arguments{
  \item{data}{data set to be used}
}
