\name{.vplayout}
\alias{.vplayout}
\title{Viewport Layout Wrapper...}
\usage{.vplayout(x, y)}
\description{Viewport Layout Wrapper
A wrapper function to set the viewport}
\keyword{internal}
\author{Hadley Wickham \email{h.wickham@gmail.com}}
\arguments{\item{x}{row position}
\item{y}{coloumn position}}
