% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs.R
\name{add_and_overwrite_aes}
\alias{add_and_overwrite_aes}
\title{Add new aes}
\usage{
add_and_overwrite_aes(current, new)
}
\value{
aes_ output
}
\description{
Add new aesthetics to a previous aes.
}
\examples{
data(diamonds, package="ggplot2")
diamonds.samp <- diamonds[sample(1:dim(diamonds)[1], 1000), ]
pm <- ggpairs(diamonds.samp, columns = 5:7,
  mapping = ggplot2::aes(color = color),
  upper = list(continuous = "cor", mapping = ggplot2::aes_string(color = "clarity")),
  lower = list(continuous = "cor", mapping = ggplot2::aes_string(color = "cut")),
  title = "Diamonds Sample"
)
str(pm)

}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{internal}
