% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglyph.R
\name{add_ref_boxes}
\alias{add_ref_boxes}
\title{Add reference boxes around each cell of the glyphmap.}
\usage{
add_ref_boxes(
  data,
  var_fill = NULL,
  color = "white",
  size = 0.5,
  fill = NA,
  ...
)
}
\arguments{
\item{data}{A glyphmap structure.}

\item{var_fill}{Variable name to use to set the fill color}

\item{color}{Set the color to draw in, default is "white"}

\item{size}{Set the line size, default is 0.5}

\item{fill}{fill value used if \code{var_fill} is \code{NULL}}

\item{...}{other arguments passed onto \code{\link[ggplot2:geom_rect]{ggplot2::geom_rect()}}}
}
\description{
Add reference boxes around each cell of the glyphmap.
}
