% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglyph.R
\name{add_ref_lines}
\alias{add_ref_lines}
\title{Add reference lines for each cell of the glyphmap.}
\usage{
add_ref_lines(data, color = "white", size = 1.5, ...)
}
\arguments{
\item{data}{A glyphmap structure.}

\item{color}{Set the color to draw in, default is "white"}

\item{size}{Set the line size, default is 1.5}

\item{...}{other arguments passed onto \code{\link[ggplot2:geom_line]{ggplot2::geom_line()}}}
}
\description{
Add reference lines for each cell of the glyphmap.
}
