% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnostic.R
\name{brew_colors}
\alias{brew_colors}
\title{RColorBrewer Set1 colors}
\usage{
brew_colors(col)
}
\arguments{
\item{col}{standard color name used to retrieve hex color value}
}
\description{
RColorBrewer Set1 colors
}
