% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggparcoord.R
\name{column_is_character}
\alias{column_is_character}
\alias{column_is_factor}
\title{Get vector of variable types from data frame}
\usage{
column_is_character(df)

column_is_factor(df)
}
\arguments{
\item{df}{data frame to extract variable types from}
}
\value{
character vector with variable types, with names corresponding to
  the variable names from df
}
\description{
Get vector of variable types from data frame
}
\author{
Barret Schloerke
}
\keyword{internal}

