% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{eval_data_col}
\alias{eval_data_col}
\title{Evaluate data column}
\usage{
eval_data_col(data, aes_col)
}
\arguments{
\item{data}{data set to evaluate the data with}

\item{aes_col}{Single value from an \code{ggplot2::\link[ggplot2]{aes}(...)} object}
}
\value{
Aes mapping with the x and y values switched
}
\description{
Evaluate data column
}
\examples{
mapping <- ggplot2::aes(Petal.Length)
eval_data_col(iris, mapping$x)
}
