\name{eval_ggpair}
\alias{eval_ggpair}
\title{Evaluate a GGally Function...}
\usage{eval_ggpair(txt, ggally_data)}
\description{Evaluate a GGally Function}
\details{Evaluate and GGally function with data, mapping, and parameters}
\keyword{internal}
\arguments{\item{txt}{text that should be evaluated to create a plot}
\item{ggally_data}{data that should be used when evaluating the text}}
