\name{find_plot_type}
\alias{find_plot_type}
\title{Find Plot Types
Retrieves the type of plot for the specific columns}
\usage{
  find_plot_type(data, col1, col2)
}
\arguments{
  \item{data}{data set to be used}

  \item{col1}{x column}

  \item{col2}{y column}
}
\description{
  Find Plot Types Retrieves the type of plot for the
  specific columns
}
\author{
  Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{internal}

