% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggpairs.r
\name{first_non_null}
\alias{first_non_null}
\title{Get first non null value}
\usage{
first_non_null(...)
}
\arguments{
\item{...}{args to be checked}
}
\description{
Get first non null value
}
\examples{
p <- qplot(1:10, 1:10) + theme(plot.title = element_text(size = 13))
first_non_null(get_theme_element(p, "plot.title", "size"), 15)
first_non_null(get_theme_element(p, "plot.title", "BAD"), 15)
}
\keyword{internal}

