\name{get.VarTypes}
\alias{get.VarTypes}
\title{Get vector of variable types from data frame}
\usage{
  get.VarTypes(df)

  get.VarTypes(df)
}
\arguments{
  \item{df}{data frame to extract variable types from}

  \item{df}{data frame to extract variable types from}
}
\value{
  character vector with variable types, with names
  corresponding to the variable names from df

  character vector with variable types, with names
  corresponding to the variable names from df
}
\description{
  Get vector of variable types from data frame

  Get vector of variable types from data frame
}
\author{
  Jason Crowley \email{crowley.jason.s@gmail.com}

  Jason Crowley \email{crowley.jason.s@gmail.com}
}

