% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find-combo.r
\name{get_plot_type}
\alias{get_plot_type}
\title{Get Plot Type}
\usage{
get_plot_type(y1_type, y2_type)
}
\arguments{
\item{y1_type}{x type.  Either numeric or category}

\item{y2_type}{y type.  Either numeric or category}
}
\description{
Retrieves the type of plot for specific info
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{internal}

