% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggmatrix_print.R
\name{get_theme_element}
\alias{get_theme_element}
\title{Get theme element}
\usage{
get_theme_element(p, element, elementKey)
}
\arguments{
\item{p}{ggplot2 object}

\item{element}{first key}

\item{elementKey}{key within element object}
}
\description{
Get the info from theme or a default value
}
\keyword{internal}

