\name{get_x_axis_labels}
\alias{get_x_axis_labels}
\title{Get x axis labels
retrieve it from the plot object directly}
\usage{
  get_x_axis_labels(p, xRange)
}
\arguments{
  \item{p}{plot object}

  \item{xRange}{range of x values}
}
\description{
  Get x axis labels retrieve it from the plot object
  directly
}
\keyword{internal}

