% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GGally-package.R
\docType{package}
\name{GGally-package}
\alias{GGally}
\alias{GGally-package}
\title{GGally: Extension to 'ggplot2'}
\description{
The R package 'ggplot2' is a plotting system based on the grammar of graphics. 'GGally' extends 'ggplot2' by adding several functions to reduce the complexity of combining geometric objects with transformed data. Some of these functions include a pairwise plot matrix, a two group pairwise plot matrix, a parallel coordinates plot, a survival plot, and several functions to plot networks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ggobi.github.io/ggally/}
  \item \url{https://github.com/ggobi/ggally}
  \item Report bugs at \url{https://github.com/ggobi/ggally/issues}
}

}
\author{
\strong{Maintainer}: Barret Schloerke \email{schloerke@gmail.com} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})

Authors:
\itemize{
  \item Barret Schloerke \email{schloerke@gmail.com} (\href{https://orcid.org/0000-0001-9986-114X}{ORCID})
  \item Di Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID}) [thesis advisor]
  \item Joseph Larmarange \email{joseph@larmarange.net} (\href{https://orcid.org/0000-0001-7097-700X}{ORCID})
  \item Francois Briatte \email{f.briatte@gmail.com}
  \item Moritz Marbach \email{mmarbach@mail.uni-mannheim.de}
  \item Edwin Thoen \email{edwinthoen@gmail.com}
  \item Amos Elberg \email{amos.elberg@gmail.com}
  \item Jason Crowley \email{crowley.jason.s@gmail.com}
}

Other contributors:
\itemize{
  \item Ott Toomet \email{otoomet@gmail.com} [contributor]
  \item Heike Hofmann \email{hhofmann4@unl.edu} (\href{https://orcid.org/0000-0001-6216-5183}{ORCID}) [thesis advisor]
  \item Hadley Wickham \email{h.wickham@gmail.com} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
