% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_prop.R
\name{ggally_colbar}
\alias{ggally_colbar}
\alias{ggally_rowbar}
\title{Column and Row bar plots}
\usage{
ggally_colbar(
  data,
  mapping,
  label_format = scales::label_percent(accuracy = 0.1),
  remove_y_axis = TRUE,
  ...,
  geom_bar_args = NULL
)

ggally_rowbar(
  data,
  mapping,
  label_format = scales::label_percent(accuracy = 0.1),
  remove_x_axis = TRUE,
  ...,
  geom_bar_args = NULL
)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{label_format}{formatter function for displaying proportions}

\item{remove_y_axis}{should y-axis be removed?}

\item{...}{other arguments passed to \code{\link[ggplot2]{geom_text}(...)}}

\item{geom_bar_args}{other arguments passed to \code{\link[ggplot2]{geom_bar}(...)}}

\item{remove_x_axis}{should x-axis be removed?}
}
\description{
Plot column or row percentage using bar plots.
}
\examples{
data(tips, package = "reshape")
ggally_colbar(tips, mapping = aes(x = smoker, y = sex))
ggally_rowbar(tips, mapping = aes(x = smoker, y = sex))

# change labels' size
ggally_colbar(tips, mapping = aes(x = smoker, y = sex), size = 8)

# change labels' colour and use bold
ggally_colbar(tips, mapping = aes(x = smoker, y = sex),
              colour = "white", fontface = "bold")

# custom bar width
ggally_colbar(tips, mapping = aes(x = smoker, y = sex), geom_bar_args = list(width = .5))

# change format of labels
ggally_colbar(tips, mapping = aes(x = smoker, y = sex),
              label_format = scales::label_percent(accuracy = .01, decimal.mark = ","))

ggduo(
  data = as.data.frame(Titanic),
  mapping = aes(weight = Freq),
  columnsX = "Survived",
  columnsY = c("Sex", "Class", "Age"),
  types = list(discrete = "rowbar"),
  legend = 1
)
}
\author{
Joseph Larmarange \email{joseph@larmarange.net}
}
\keyword{hplot}
