% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ggally_cor_v1_5}
\alias{ggally_cor_v1_5}
\title{Correlation value plot}
\usage{
ggally_cor_v1_5(
  data,
  mapping,
  alignPercent = 0.6,
  method = "pearson",
  use = "complete.obs",
  corAlignPercent = NULL,
  corMethod = NULL,
  corUse = NULL,
  displayGrid = TRUE,
  ...
)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{alignPercent}{right align position of numbers. Default is 60 percent across the horizontal}

\item{method}{\code{method} supplied to cor function}

\item{use}{\code{use} supplied to cor function}

\item{corAlignPercent}{deprecated. Use parameter \code{alignPercent}}

\item{corMethod}{deprecated. Use parameter \code{method}}

\item{corUse}{deprecated. Use parameter \code{use}}

\item{displayGrid}{if TRUE, display aligned panel gridlines}

\item{...}{other arguments being supplied to geom_text}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Estimate correlation from the given data.

This function is deprecated and will be removed in future releases.

See \code{\link{ggally_cor}}.
}
\examples{
# Small function to display plots only if it's interactive
p_ <- GGally::print_if_interactive

data(tips)
p_(ggally_cor_v1_5(tips, mapping = ggplot2::aes(total_bill, tip)))

# display with no grid
p_(ggally_cor_v1_5(
  tips,
  mapping = ggplot2::aes(total_bill, tip),
  displayGrid = FALSE
))

# change text attributes
p_(ggally_cor_v1_5(
  tips,
  mapping = ggplot2::aes(x = total_bill, y = tip),
  size = 15,
  colour = I("red")
))

# split by a variable
p_(ggally_cor_v1_5(
  tips,
  mapping = ggplot2::aes(total_bill, tip, color = sex),
  size = 5
))
}
\seealso{
\code{\link{ggally_cor}}
}
\author{
Barret Schloerke
}
\keyword{hplot}
\keyword{internal}
