% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{ggally_density_ridges}
\alias{ggally_density_ridges}
\title{Create a ridgeline plot.}
\usage{
ggally_density_ridges(data, mapping, ...)
}
\arguments{
\item{data}{The data set used.}

\item{mapping}{The aesthetics used.}

\item{...}{Other parameters passed on to \code{\link[ggridges]{geom_density_ridges}}.}
}
\description{
`ggally_density_ridges` arranges multiple density plots in a staggered fashion, as in the cover of the famous Joy Division album Unknown Pleasures.
}
\examples{
  data(iris)
  ggpairs(iris, lower = list("combo" = ggally_density_ridges))
  ggpairs(iris, lower = list("combo" = ggally_density_ridges, mapping = ggplot2::aes(color = Species)))
  ggpairs(iris, lower = list("combo" = wrap(ggally_density_ridges, rel_min_height = 0.01), mapping = ggplot2::aes(color = Species)))

}
\author{
Jens Preussner \email{jens.preussner@mpi-bn.mpg.de}
}
