% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{ggally_dot_and_box}
\alias{ggally_dot_and_box}
\title{Plots either Box Plot or Dot Plots}
\usage{
ggally_dot_and_box(data, mapping, ..., boxPlot = TRUE)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{...}{parameters passed to either geom_jitter or geom_boxplot}

\item{boxPlot}{boolean to decide to plot either box plots (TRUE) or dot plots (FALSE)}
}
\description{
Place box plots or dot plots on the graph
}
\examples{
 data(tips, package = "reshape")
 ggally_dot_and_box(
   tips,
   mapping = ggplot2::aes(x = total_bill, y = sex, color = sex),
   boxPlot = TRUE
 )
 ggally_dot_and_box(
   tips,
   mapping = ggplot2::aes(x = total_bill, y = sex, color = sex),
   boxPlot = FALSE
 )
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}
