\name{ggally_dotAndBox}
\alias{ggally_dotAndBox}
\title{Plots either Box Plot or Dot Plots...}
\usage{ggally_dotAndBox(data, ..., boxPlot=TRUE)}

\description{
  Plots either Box Plot or Dot Plots Place box plots or dot plots on
  the graph
}
\author{Barret Schloerke \email{bigbear@iastate.edu}, Haesung Kim \email{hae0510@iastate.edu}, Di Cook \email{dicook@iastate.edu} and Heike Hofmann \email{hofmann@iastate.edu}}
\keyword{hplot}
\arguments{
  \item{data}{data set using}
  \item{...}{parameters passed to aes_string(...)}
  \item{boxPlot}{boolean to decide to plot either box plots (TRUE) or dot plots (FALSE)}
}
\examples{example(ggally_box)
example(ggally_dot)}
