% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{ggally_facethist}
\alias{ggally_facethist}
\title{Faceted histogram}
\usage{
ggally_facethist(data, mapping, ...)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{...}{parameters sent to stat_bin()}
}
\description{
Display subsetted histograms of the data in different panels.
}
\examples{
 data(tips, package = "reshape")
 ggally_facethist(tips, mapping = ggplot2::aes(x = tip, y = sex))
 ggally_facethist(tips, mapping = ggplot2::aes_string(x = "tip", y = "sex"), binwidth = 0.1)
}
\author{
Barret Schloerke
}
\keyword{hplot}
