% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{ggally_na}
\alias{ggally_na}
\alias{ggally_naDiag}
\title{NA plot}
\usage{
ggally_na(..., size = 10)

ggally_naDiag(...)
}
\arguments{
\item{...}{other arguments ignored}

\item{size}{size of the geom_text 'NA'}
}
\description{
Draws a large \code{NA} in the middle of the plotting area.  This plot is useful when all X or Y data is \code{NA}
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}

