% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_cross.R
\name{ggally_table}
\alias{ggally_table}
\title{Display a table of the number of observations}
\usage{
ggally_table(data, mapping, ..., geom_tile_args = NULL)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{...}{other arguments passed to \code{\link[ggplot2]{geom_text}(...)}}

\item{geom_tile_args}{other arguments passed to \code{\link[ggplot2]{geom_tile}(...)}}
}
\description{
Plot the number of observations as a table. Other statistics computed
by \code{\link{stat_cross}}  could be used (see examples).
}
\note{
The \strong{colour} aesthetic is taken into account only if equal to
  \strong{x} or \strong{y}.
}
\examples{
data(tips, package = "reshape")
ggally_table(tips, mapping = aes(x = smoker, y = sex))
ggally_table(tips, mapping = aes(x = smoker, y = sex, colour = smoker))
ggally_table(tips, mapping = aes(x = smoker, y = sex, colour = day))
ggally_tableDiag(tips, mapping = aes(x = smoker))

# Custom label size and color
ggally_table(tips, mapping = aes(x = smoker, y = sex), size = 16, color = "red")

# Display column proportions
ggally_table(tips, mapping = aes(x = day, y = sex, label = scales::percent(after_stat(col.prop))))

# Draw table cells
ggally_table(
  tips,
  mapping = aes(x = smoker, y = sex),
  geom_tile_args = list(colour = "black", fill = "white")
)

# Use standardized residuals to fill table cells
ggally_table(
  as.data.frame(Titanic),
  mapping = aes(
    x = Class, y = Survived, weight = Freq,
    fill = after_stat(stdres),
    label = scales::percent(after_stat(col.prop), accuracy = .1)
  ),
  geom_tile_args = list(colour = "black")
) +
scale_fill_steps2(breaks = c(-3, -2, 2, 3), show.limits = TRUE)
}
\author{
Joseph Larmarange \email{joseph@larmarange.net}
}
\keyword{hplot}
