% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggfacet.R
\name{ggfacet}
\alias{ggfacet}
\title{ggfacet - single ggplot2 plot matrix with facet_grid}
\usage{
ggfacet(data, mapping = NULL, columnsX = 1:ncol(data),
  columnsY = 1:ncol(data), fn = "points", ...,
  columnLabelsX = names(data[columnsX]),
  columnLabelsY = names(data[columnsY]), xlab = NULL, ylab = NULL,
  title = NULL, scales = "free")
}
\arguments{
\item{data}{data.frame that contains all columns to be displayed.  This data will be melted before being passed into the function \code{fn}}

\item{mapping}{aesthetic mapping (besides \code{x} and \code{y}).  See \code{\link[ggplot2]{aes}()}}

\item{columnsX}{columns to be displayed in the plot matrix}

\item{columnsY}{rows to be displayed in the plot matrix}

\item{fn}{function to be executed. Similar to \code{\link{ggpairs}} and \code{\link{ggduo}}, the function may either be a string identifier or a real function that \code{\link{wrap}} understands.}

\item{...}{extra arguments passed directly to \code{fn}}

\item{columnLabelsX, columnLabelsY}{column and row labels to display in the plot matrix}

\item{xlab, ylab, title}{plot matrix labels}

\item{scales}{parameter supplied to \code{ggplot2::\link[ggplot2]{facet_grid}}. Default behavior is \code{"free"}}
}
\description{
ggfacet - single ggplot2 plot matrix with facet_grid
}
\examples{
# Small function to display plots only if it's interactive
p_ <- GGally::print_if_interactive
data(mtcars)

small_mtcars <- mtcars[c(1,3:6)]

# Basic plot of all column pair combinations
p_(ggfacet(small_mtcars))

# Change scales to 'fixed' to display each plot on the same scale
p_(ggfacet(small_mtcars, scales = "fixed"))

# Change to ggally_smooth as plotting function
p_(ggfacet(small_mtcars, fn = 'smooth'))
# Same output
p_(ggfacet(small_mtcars, fn = ggally_smooth))

# Change ggally_smooth's smoothing method to 'loess'
p_(ggfacet(small_mtcars, fn = "smooth", method = "loess"))

# Look at different response variables in mtcars
p_(ggfacet(mtcars, columnsX = 3:6, columnsY = c("mpg", "qsec"), fn = "smooth", method = "loess"))
}

