% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{ggfluctuation2}
\alias{ggfluctuation2}
\title{Fluctuation plot - deprecated}
\usage{
ggfluctuation2(table_data, floor = 0, ceiling = max(table_data$freq, na.rm =
  TRUE))
}
\arguments{
\item{table_data}{a table of values, or a data frame with three columns, the last column being frequency}

\item{floor}{don't display cells smaller than this value}

\item{ceiling}{max value to compare to}
}
\description{
Fluctuation plot - deprecated
}
\examples{
data(tips, package = "reshape")
ggfluctuation2(table(tips$sex, tips$day))
ggfluctuation2(table(tips[, c("sex", "day")]))
}
\author{
Hadley Wickham \email{h.wickham@gmail.com}, Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}

