% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggmatrix.R
\name{ggmatrix}
\alias{ggmatrix}
\title{ggpairs - A ggplot2 Matrix}
\usage{
ggmatrix(plots, nrow, ncol, xAxisLabels = rep("", length(plots)),
  yAxisLabels = rep("", length(plots)), axisLabels = "show", title = NULL,
  byrow = TRUE, verbose = FALSE, data = NULL, gg = NULL,
  legends = FALSE)
}
\arguments{
\item{plots}{list of plots to be put into matrix}

\item{nrow,ncol}{number of rows and columns}

\item{xAxisLabels,yAxisLabels,title}{labels for plot}

\item{axisLabels}{either "show" to display axisLabels, "internal" for labels in the diagonal plots, or "none" for no axis labels}

\item{byrow}{boolean that determines whether the plots should be ordered by row or by column}

\item{verbose}{boolean to determine the printing of "Plot #1, Plot #2...."}

\item{data}{data set using. This is the data to be used in place of 'ggally_data' if the plot is a string to be evaluated at print time}

\item{gg}{ggplot2 theme objects to be applied to every plot}

\item{legends}{boolean to determine the printing of the legend in each plot. Not recommended.}
}
\description{
Make a matrix of ggplot2 plots
}
\examples{
plotList <- list()
for (i in 1:6) {
  plotList[[i]] <- ggally_text(paste("Plot #", i, sep = ""))
}
a <- ggmatrix(
  plotList,
  2, 3,
  c("A", "B", "C"),
  c("D", "E"),
  byrow = TRUE
)
#a

a <- ggmatrix(
  plotList,
  2, 3,
  c("A", "B", "C"),
  c("D", "E"),
  byrow = FALSE
)
#a
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}

