% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmatrix_print_new.R
\name{ggprint}
\alias{ggprint}
\title{Print ggpair object}
\usage{
ggprint(x, ..., progress = interactive(), progress_wait = 5)
}
\arguments{
\item{x}{ggpair object to be plotted}

\item{...}{ignored}
}
\description{
Specialized method to print the ggpair object-
}
\examples{
 data(tips, package = "reshape")
 pm <- ggpairs(tips, c(1,3,2), mapping = ggplot2::aes_string(color = "sex"))
 ggprint(pm) # calls print(pm), which calls ggprint.ggmatrix(pm)
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}

