% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglyph.R
\name{glyphplot}
\alias{glyphplot}
\alias{is.glyphplot}
\alias{[.glyphplot}
\alias{print.glyphplot}
\title{Glyph plot class}
\usage{
glyphplot(data, width, height, polar, x_major, y_major)

is.glyphplot(x)

\method{[}{glyphplot}(x, ...)

\method{print}{glyphplot}(x, ...)
}
\arguments{
\item{data}{A data frame containing variables named in \code{x_major},
\code{x_minor}, \code{y_major} and \code{y_minor}.}

\item{height, width}{The height and width of each glyph. Defaults to 95\% of
the \code{\link[ggplot2]{resolution}} of the data. Specify the width
absolutely by supplying a numeric vector of length 1, or relative to the}

\item{polar}{A logical of length 1, specifying whether the glyphs should
be drawn in polar coordinates.  Defaults to \code{FALSE}.}

\item{x_major, y_major}{The name of the variable (as a
string) for the major x and y axes.  Together, the}

\item{x}{glyphplot to be printed}

\item{...}{ignored}
}
\description{
Glyph plot class
}
\author{
Di Cook \email{dicook@monash.edu}, Heike Hofmann, Hadley Wickham
}
