% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs_add.R
\name{is_ggmatrix}
\alias{is_ggmatrix}
\title{Check if an object is a ggmatrix}
\usage{
is_ggmatrix(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
Logical value indicating if the object is a \code{ggmatrix}
}
\description{
Check if an object is a ggmatrix
}
\examples{
is_ggmatrix(ggpairs(mtcars))
is_ggmatrix(ggplot2::ggplot())
}
