% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{is_horizontal}
\alias{is_horizontal}
\alias{is_character_column}
\title{Check if plot is horizontal}
\usage{
is_horizontal(data, mapping, val = "y")

is_character_column(data, mapping, val = "y")
}
\arguments{
\item{data}{data used in ggplot2 plot}

\item{mapping}{ggplot2 \code{aes()} mapping}

\item{val}{key to retrieve from \code{mapping}}
}
\value{
Boolean determining if the data is a character-like data
}
\description{
Check if plot is horizontal
}
\examples{
is_horizontal(iris, ggplot2::aes(Sepal.Length, Species)) # TRUE
is_horizontal(iris, ggplot2::aes(Sepal.Length, Species), "x") # FALSE
is_horizontal(iris, ggplot2::aes(Sepal.Length, Sepal.Width)) # FALSE
}
