% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{keep_colour_if_in_x_or_y}
\alias{keep_colour_if_in_x_or_y}
\title{Keep colour mapping only if same as x or y}
\usage{
keep_colour_if_in_x_or_y(mapping)
}
\arguments{
\item{mapping}{output of \code{ggplot2::\link[ggplot2]{aes}(...)}}
}
\value{
Aes mapping with colour mapping kept only if equal to
mapping to x or to y.
}
\description{
Keep colour mapping only if same as x or y
}
\examples{
mapping <- aes(x = sex, y = age, colour = sex)
keep_colour_if_in_x_or_y(mapping)

mapping <- aes(x = sex, y = age, colour = region)
keep_colour_if_in_x_or_y(mapping)
}
