% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lowertriangle}
\alias{lowertriangle}
\title{lowertriangle - rearrange dataset as the preparation of ggscatmat function}
\usage{
lowertriangle(data, columns = 1:ncol(data), color = NULL)
}
\arguments{
\item{data}{a data matrix. Should contain numerical (continuous) data.}

\item{columns}{an option to choose the column to be used in the raw dataset. Defaults to \code{1:ncol(data)}}

\item{color}{an option to choose a factor variable to be grouped with. Defaults to \code{(NULL)}}
}
\description{
function for making the melted dataset used to plot the lowertriangle scatterplots.
}
\examples{
data(flea)
head(lowertriangle(flea, columns= 2:4))
head(lowertriangle(flea))
head(lowertriangle(flea, color="species"))
}
\author{
Mengjia Ni, Di Cook \email{dicook@iastate.edu}
}

