\name{make_ggpair_text}
\alias{make_ggpair_text}
\title{Generate GGally Function Text...}
\usage{make_ggpair_text(func, mapping, params, printInfo=FALSE)}
\description{Generate GGally Function Text}
\details{Generate GGally function text with data, mapping, and parameters}
\keyword{internal}
\arguments{\item{func}{identifier string in function name}
\item{mapping}{mapping supplied to the function}
\item{params}{parameters applied to the geom in the function}
\item{printInfo}{boolean to determine whether or not the executed function should be printed}}
