\name{mapping_color_fill}
\alias{mapping_color_fill}
\title{Aesthetic Mapping Color Fill
Replace the fill with the color and make color NULL}
\usage{
  mapping_color_fill(current)
}
\arguments{
  \item{current}{the current aesthetics}
}
\description{
  Aesthetic Mapping Color Fill Replace the fill with the
  color and make color NULL
}
\keyword{internal}

