% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs.R
\name{mapping_color_to_fill}
\alias{mapping_color_to_fill}
\title{Aesthetic Mapping Color Fill}
\usage{
mapping_color_to_fill(current)
}
\arguments{
\item{current}{the current aesthetics}
}
\description{
Replace the fill with the color and make color NULL.
}
\keyword{internal}
