% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{mapping_string}
\alias{mapping_string}
\title{Aes name}
\usage{
mapping_string(aes_col)
}
\arguments{
\item{aes_col}{Single value from \code{ggplot2::\link[ggplot2]{aes}(...)}}
}
\value{
character string
}
\description{
Aes name
}
\examples{
mapping <- ggplot2::aes(Petal.Length)
mapping_string(mapping$x)
}
