% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{mapping_swap_x_y}
\alias{mapping_swap_x_y}
\title{Swap x and y mapping}
\usage{
mapping_swap_x_y(mapping)
}
\arguments{
\item{mapping}{output of \code{ggplot2::\link[ggplot2]{aes}(...)}}
}
\value{
Aes mapping with the x and y values switched
}
\description{
Swap x and y mapping
}
\examples{
mapping <- ggplot2::aes(Petal.Length, Sepal.Width)
mapping
mapping_swap_x_y(mapping)
}
