% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nba_ppg_2008.R
\docType{data}
\name{nba_ppg_2008}
\alias{nba_ppg_2008}
\title{NBA Player Statistics for 2008-2009 Season}
\format{
A data frame with 50 rows and 21 variables
}
\source{
Data originally collected from FlowingData tutorial
}
\usage{
data(nba_ppg_2008)
}
\description{
This dataset contains performance statistics for NBA players from the 2008-2009 season.
The data includes top-performing players with their scoring averages and various
basketball performance metrics.
}
\details{
The dataset contains the following variables:
\itemize{
\item Name - Player name
\item G - Games played
\item MIN - Minutes per game
\item PTS - Points per game
\item FGM - Field goals made per game
\item FGA - Field goal attempts per game
\item FGP - Field goal percentage
\item FTM - Free throws made per game
\item FTA - Free throw attempts per game
\item FTP - Free throw percentage
\item X3PM - Three-point field goals made per game
\item X3PA - Three-point field goal attempts per game
\item X3PP - Three-point field goal percentage
\item ORB - Offensive rebounds per game
\item DRB - Defensive rebounds per game
\item TRB - Total rebounds per game
\item AST - Assists per game
\item STL - Steals per game
\item BLK - Blocks per game
\item TO - Turnovers per game
\item PF - Personal fouls per game
}
}
\references{
FlowingData (2010) How to Make a Heatmap - a Quick and Easy Solution.
\url{https://flowingdata.com/2010/01/21/how-to-make-a-heatmap-a-quick-and-easy-solution/}
}
\keyword{datasets}
