\name{plot_types}
\alias{plot_types}
\title{Plot Types
Retrieves the type of plot that should be used for all combinations}
\usage{
  plot_types(data)
}
\arguments{
  \item{data}{data set to be used}
}
\description{
  Plot Types Retrieves the type of plot that should be used
  for all combinations
}
\author{
  Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{internal}

