\name{print.ggpairs}
\alias{print.ggpairs}
\title{Print ggpair object...}
\usage{\method{print}{ggpairs} (x, ...)}
\description{Print ggpair object
Specialized method to print the ggapir object}
\keyword{internal}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\arguments{\item{x}{ggpair object to be plotted}
\item{...}{not used}}
\examples{ggpairs(iris[,3:5])}
