% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print_if_interactive}
\alias{print_if_interactive}
\title{Print if not CRAN}
\usage{
print_if_interactive(p)
}
\arguments{
\item{p}{plot to be displayed}
}
\description{
Small function to print a plot if the R session is interactive or in a travis build
}
