\name{putPlot}
\alias{putPlot}
\title{Put Plot...}
\usage{putPlot(rowFromTop,
    columnFromLeft, p, axes=TRUE)}
\description{Put Plot
Function to place your own plot in the layout}
\keyword{hplot}
\author{Barret Schloerke \email{bigbear@iastate.edu} and Haesung Kim \email{hae0510@iastate.edu}}
\arguments{\item{rowFromTop}{row from the top}
\item{columnFromLeft}{column from the left}
\item{p}{ggplot object to be placed}
\item{axes}{Boolean to tell wether or not to print the x and y labels, axis text, and legends}}
\examples{example(ggpairs)}
