% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglyph.R
\name{rel}
\alias{is.rel}
\alias{print.rel}
\alias{rel}
\title{Relative dimensions}
\usage{
rel(x)

\method{print}{rel}(x, ...)

is.rel(x)
}
\arguments{
\item{x}{numeric value between 0 and 1}

\item{...}{ignored}
}
\description{
Relative dimensions
}
\author{
Di Cook \email{dicook@iastate.edu}
}

