% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{require_namespaces}
\alias{require_namespaces}
\title{Loads package namespaces}
\usage{
require_namespaces(pkgs)
}
\arguments{
\item{pkgs}{vector of character values}
}
\description{
Loads package namespaces or yells at user... loudly
}
\keyword{internal}
