% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{require_pkgs}
\alias{require_pkgs}
\title{Require packages}
\usage{
require_pkgs(pkgs)
}
\arguments{
\item{pkgs}{vector of character values}
}
\description{
Requires packages or yells at user... loudly
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{internal}
