% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglyph.R
\name{range01}
\alias{range01}
\alias{max1}
\alias{mean0}
\alias{min0}
\alias{rescale01}
\alias{rescale11}
\title{Rescaling functions}
\usage{
range01(x)

max1(x)

mean0(x)

min0(x)

rescale01(x, xlim = NULL)

rescale11(x, xlim = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{xlim}{value used in \code{range}}
}
\description{
Rescaling functions
}
\section{Functions}{
\itemize{
\item \code{range01()}: Rescale all values to \verb{[0, 1]}

\item \code{max1()}: Divide by the maximum value

\item \code{mean0()}: Subtract the mean value

\item \code{min0()}: Subtract the minimum value

\item \code{rescale01()}: Rescale all values given a range. If no range is
provided, the min and max values are used. By default, this will result in
values from \verb{[0, 1]}.

\item \code{rescale11()}: Rescale all values given a range. If no range is
provided, the min and max values are used. By default, this will result in
values from \verb{[-1, 1]}.

}}
