% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparcoord.R
\name{scag_order}
\alias{scag_order}
\title{Find order of variables}
\usage{
scag_order(scag, vars, measure)
}
\arguments{
\item{scag}{\code{scagnostics} object}

\item{vars}{character vector of the variables to be ordered}

\item{measure}{scagnostics measure to order according to}
}
\value{
character vector of variable ordered according to the given
  scagnostic measure
}
\description{
Find order of variables based on a specified scagnostic measure
by maximizing the index values of that measure along the path.
}
\author{
Barret Schloerke
}
