% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs.R
\name{set_to_blank_list_if_blank}
\alias{set_to_blank_list_if_blank}
\title{Aesthetic Mapping Color Fill}
\usage{
set_to_blank_list_if_blank(val, combo = TRUE)
}
\arguments{
\item{current}{the current aesthetics}
}
\description{
Replace the fill with the color and make color NULL.
}
\keyword{internal}

