\name{skewness}
\alias{skewness}
\title{Sample skewness}
\usage{
  skewness(x)

  skewness(x)
}
\arguments{
  \item{x}{numeric vector}

  \item{x}{numeric vector}
}
\value{
  sample skewness of \code{x}

  sample skewness of \code{x}
}
\description{
  Calculate the sample skewness of a vector while ignoring
  missing values.

  Calculate the sample skewness of a vector while ignoring
  missing values.
}
\author{
  Jason Crowley \email{crowley.jason.s@gmail.com}

  Jason Crowley \email{crowley.jason.s@gmail.com}
}

