% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_at_risk_table.R
\name{surv_at_risk_table}
\alias{surv_at_risk_table}
\title{Number at risk table for survival graph}
\usage{
surv_at_risk_table(s, xlim = c(0, max(s$time)),
  at.risk.interval = diff(xlim)/10, surv.col = NULL)
}
\arguments{
\item{s}{an object of class \code{\link[survival]{survfit}}}

\item{xlim}{limits of time parameter (defaults to range of \code{s$time})}

\item{at.risk.interval}{interval at which to display number at risk (defaults to \code{diff(xlim) / 10})}

\item{surv.col}{colours for}
}
\value{
custom annotation for adding to a \code{\link[ggplot2]{ggplot}}
}
\description{
Number at risk table for survival graph
}
\details{
This function produces a number at risk table from a \code{\link[survival]{survfit}} object. It can
  be placed in the margin of a \code{\link{ggsurv}} plot using a customAnnotation (although
  alignment may be a little tricky) or combined into a composite plot using, for example,
  \code{\link[cowplot]{plot_grid}}.
}
\examples{
library(survival)
fit <- survfit(Surv(time, status) ~ x, aml)
surv_at_risk_table(fit, at.risk.interval = 10)

if (require(cowplot)) {
  plot_grid(
    ggsurv(fit),
    surv_at_risk_table(fit, at.risk.interval = 10),
    nrow = 2,
    rel_heights = c(10, 2),
    align = "v"
  )

  # Note that positioning of legends will need to be applied both to the survival
  # plot and the number at risk table to ensure alignment still works
  legend_theme <- theme(legend.position = "none")
  plot_grid(
    ggsurv(fit) + legend_theme,
    surv_at_risk_table(fit, at.risk.interval = 10, coloured = TRUE) + legend_theme,
    nrow = 2,
    rel_heights = c(10, 2),
    align = "v"
  )
}
}
