% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tips.R
\docType{data}
\name{tips}
\alias{tips}
\title{Tipping data}
\format{
A data frame with 244 rows and 7 variables
}
\usage{
tips
}
\description{
One waiter recorded information about each tip he received over a
period of a few months working in one restaurant. He collected several
variables:
}
\details{
\itemize{
\item tip in dollars,
\item bill in dollars,
\item sex of the bill payer,
\item whether there were smokers in the party,
\item day of the week,
\item time of day,
\item size of the party.
}

In all he recorded 244 tips. The data was reported in a collection of
case studies for business statistics (Bryant & Smith 1995).
}
\references{
Bryant, P. G. and Smith, M (1995) \emph{Practical Data
Analysis: Case Studies in Business Statistics}.  Homewood, IL: Richard D.
Irwin Publishing:
}
\keyword{datasets}
