% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-twitter_spambots.R
\docType{data}
\name{twitter_spambots}
\alias{twitter_spambots}
\title{Twitter spambots}
\format{An object of class \code{network} with 120 edges and 94 vertices.}
\usage{
data(twitter_spambots)
}
\description{
A network of spambots found on Twitter as part of a data mining project.
}
\details{
Each node of the network is identified by the Twitter screen name of the
account and further carries five vertex attributes:

\itemize{
  \item location user's location, as provided by the user
  \item lat latitude, based on the user's location
  \item lon longitude, based on the user's location
  \item followers number of Twitter accounts that follow this account
  \item friends number of Twitter accounts followed by the account
}
}
\author{
Amos Elberg
}
\keyword{datasets}
