% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vig_ggally.R
\name{vig_ggally}
\alias{vig_ggally}
\title{View \pkg{GGally} vignettes}
\usage{
vig_ggally(name)
}
\arguments{
\item{name}{Vignette name to open. If no name is provided, the vignette index will be opened}
}
\description{
This function will open the directly to the vignette requested. If no \code{name} is provided, the index of all \pkg{GGally} vignettes will be opened.
}
\details{
This method allows for vignettes to be hosted remotely, reducing \pkg{GGally}'s package size, and installation time.
}
\examples{
\donttest{
# View `ggnostic` vignette
vig_ggally("ggnostic")

# View all vignettes by GGally
vig_ggally()}
}
