\name{vplayout}
\alias{vplayout}
\title{Viewport Layout Wrapper
A wrapper function to set the viewport}
\usage{
vplayout(x, y)
}
\arguments{
  \item{x}{row position}

  \item{y}{coloumn position}
}
\description{
  Viewport Layout Wrapper A wrapper function to set the
  viewport
}
\author{
  Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}
