\name{.vplayout}
\alias{.vplayout}
\title{Viewport Layout Wrapper...}
\usage{.vplayout(x, y)}
\description{Viewport Layout Wrapper
A wrapper function to set the viewport}
\keyword{internal}
\author{Barret Schloerke \email{bigbear@iastate.edu} and Haesung Kim \email{hae0510@iastate.edu}}
\arguments{\item{x}{row position}
\item{y}{coloumn position}}
