% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_ref_boxes}
\alias{add_ref_boxes}
\title{Add reference boxes around each cell of the glyphmap.}
\usage{
add_ref_boxes(data, var_fill = NULL, colour = "white", size = 0.5,
  fill = NA, ...)
}
\arguments{
\item{data}{A glyphmap structure.}

\item{var_fill}{Variable name to use to set the fill color}

\item{size}{Set the line size, default is 0.5}

\item{color}{Set the color to draw in, default is "white"}
}
\description{
Add reference boxes around each cell of the glyphmap.
}

