% Generated by roxygen2 (4.0.2): do not edit by hand
\name{add_ref_lines}
\alias{add_ref_lines}
\title{Add reference lines for each cell of the glyphmap.}
\usage{
add_ref_lines(data, colour = "white", size = 1.5, ...)
}
\arguments{
\item{data}{A glyphmap structure.}

\item{size}{Set the line size, default is 1.5}

\item{color}{Set the color to draw in, default is "white"}
}
\description{
Add reference lines for each cell of the glyphmap.
}

