% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggduo_wrappers.R
\name{broomify}
\alias{broomify}
\title{Broomify a model}
\usage{
broomify(model)
}
\arguments{
\item{model}{model to be sent to \code{broom::\link[broom]{augment}}, \code{broom::\link[broom]{glance}}, and \code{broom::\link[broom]{tidy}}}
}
\value{
augmented data frame with the glance data.frame and tidy data.frame as 'broom_glance' and 'broom_tidy' attributes respectively
}
\description{
broom::augment a model and add broom::glance and broom::tidy output as attributes
}

