\name{eval_ggpair}
\alias{eval_ggpair}
\title{Evaluate a GGally Function...}
\usage{eval_ggpair(func, data, mapping, params, printInfo=FALSE)}
\description{Evaluate a GGally Function}
\details{Evaluate and GGally function with data, mapping, and parameters}
\arguments{\item{func}{identifier string in function name}
\item{data}{data supplied to the function}
\item{mapping}{mapping supplied to the function}
\item{params}{parameters applied to the geom in the function}
\item{printInfo}{boolean to determine whether or not the executed function should be printed}}
