% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggpairs.r
\name{eval_ggpair}
\alias{eval_ggpair}
\title{Evaluate a GGally Function}
\usage{
eval_ggpair(txt, ggally_data)
}
\arguments{
\item{txt}{text that should be evaluated to create a plot}

\item{ggally_data}{data that should be used when evaluating the text}
}
\description{
Evaluate and GGally function with data, mapping, and parameters.
}
\keyword{internal}

