\name{find_plot_type}
\alias{find_plot_type}
\title{Find Plot Types...}
\usage{find_plot_type(data, col1, col2)}
\description{Find Plot Types
Retrieves the type of plot for the specific columns}
\keyword{internal}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\arguments{\item{data}{data set to be used}
\item{col1}{x column}
\item{col2}{y column}}
