% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-combo.R
\name{find_plot_type}
\alias{find_plot_type}
\title{Find Plot Types}
\usage{
find_plot_type(col1Name, col2Name, type1, type2, isAllNa, allowDiag)
}
\arguments{
\item{col1Name}{x column name}

\item{col2Name}{y column name}

\item{type1}{x column type}

\item{type2}{y column type}

\item{isAllNa}{is.na(data)}

\item{allowDiag}{allow for diag values to be returned}
}
\description{
Retrieves the type of plot for the specific columns
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
