% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find-combo.r
\name{find_plot_type}
\alias{find_plot_type}
\title{Find Plot Types}
\usage{
find_plot_type(data, col1, col2)
}
\arguments{
\item{data}{data set to be used}

\item{col1}{x column}

\item{col2}{y column}
}
\description{
Retrieves the type of plot for the specific columns
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{internal}

