\name{get.VarTypes}
\alias{get.VarTypes}
\title{Get vector of variable types from data frame...}
\usage{get.VarTypes(df)}
\description{Get vector of variable types from data frame}
\author{Jason Crowley \email{crowley.jason.s@gmail.com}}
\value{character vector with variable types, with names corresponding to
the variable names from df}
\arguments{\item{df}{data frame to extract variable types from}}
