% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gg-plots.r
\name{get_x_axis_labels}
\alias{get_x_axis_labels}
\title{Get x axis labels}
\usage{
get_x_axis_labels(p, xRange)
}
\arguments{
\item{p}{plot object}

\item{xRange}{range of x values}
}
\description{
Retrieves x axis labels from the plot object directly.
}
\keyword{internal}

