\name{getPlot}
\alias{getPlot}
\title{getPlot...}
\usage{getPlot(x, y)}
\description{getPlot
Retrieves the ggplot object at the desired location}
\keyword{hplot}
\author{Barret Schloerke \email{bigbear@iastate.edu} and Haesung Kim \email{hae0510@iastate.edu}}
\arguments{\item{x}{row from the top}
\item{y}{column from the left}}
\examples{example(ggpairs)}
