% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs_getput.R
\name{getPlot}
\alias{[.ggmatrix}
\alias{getPlot}
\title{getPlot}
\usage{
getPlot(x, i, j)

\method{[}{ggmatrix}(pm, i, j, ...)
}
\arguments{
\item{i}{row from the top}

\item{j}{column from the left}

\item{pm}{ggpair object to select from}

\item{...}{ignored}
}
\description{
Retrieves the ggplot object at the desired location.
}
\examples{
 data(tips, package = "reshape")
 plotMatrix2 <- ggpairs(tips[, 3:2], upper = list(combo = "denstrip"))
 plotMatrix2[1, 2]
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}

