% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggpairs_getput.R
\name{getPlot}
\alias{[.ggpairs}
\alias{getPlot}
\title{getPlot}
\usage{
getPlot(x, i, j)

\method{[}{ggpairs}(x, i, j, ...)
}
\arguments{
\item{x}{ggpair object to select from}

\item{i}{row from the top}

\item{j}{column from the left}

\item{...}{ignored}
}
\description{
Retrieves the ggplot object at the desired location.
}
\examples{
data(tips, package = "reshape")
 plotMatrix2 <- ggpairs(tips[,3:2], upper = list(combo = "denstrip"))
 plotMatrix2[1, 2]
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}

