\name{GGally-package}
\alias{GGally-package}
\alias{GGally}
\docType{package}
\title{
Extension to ggplot2
}
\description{
Package contains templates for different plots to be combined into a plot matrix.
}
\details{
\tabular{ll}{
Package: \tab GGally\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2010-1-20\cr
License: \tab GPL (>= 2.0)\cr
LazyData: \tab yes\cr
}
}
\author{
Barret Schloerke <bigbear@iastate.edu>, Di Cook <dicook@iastate.edu>, Heike
    Hofmann <hofmann@iastate.edu>, and Hadley Wickham <h.wickham@iastate.edu>
Maintainer: Barret Schloerke <bigbear@iastate.edu>
}
\references{
http://www.had.co.nz/ggplot2/
}
\keyword{ package }
\examples{
	example(ggpairs)
}
