% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{ggally_autopoint}
\alias{ggally_autopoint}
\alias{ggally_autopointDiag}
\title{Scatterplot for continuous and categorial variables}
\usage{
ggally_autopoint(data, mapping, ...)

ggally_autopointDiag(data, mapping, ...)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{...}{other arguments passed to \code{\link[ggforce]{geom_autopoint}(...)}}
}
\description{
Make scatterplots compatible with both continuous and catgeorical variables
using \code{\link[ggforce]{geom_autopoint}} from package \pkg{ggforce}.
}
\examples{
data(tips, package = "reshape")
ggally_autopoint(tips, mapping = aes(x = tip, y = total_bill))
ggally_autopoint(tips, mapping = aes(x = tip, y = sex))
ggally_autopoint(tips, mapping = aes(x = smoker, y = sex))
ggally_autopoint(tips, mapping = aes(x = smoker, y = sex, color = day))
ggally_autopoint(tips, mapping = aes(x = smoker, y = sex), size = 8)
ggally_autopoint(tips, mapping = aes(x = smoker, y = sex), alpha = .9)

\dontrun{
ggpairs(
  tips,
  mapping = aes(colour = sex),
  upper = list(discrete = "autopoint", combo = "autopoint", continuous = "autopoint"),
  diag = list(discrete = "autopointDiag", continuous = "autopointDiag")
)
}
}
\keyword{hplot}
