\name{ggally_barDiag}
\alias{ggally_barDiag}
\title{Plots the Bar Plots by Using Diagonal...}
\usage{ggally_barDiag(data, ...)}
\description{Plots the Bar Plots by Using Diagonal
Plots the bar plots by using Diagonal}
\author{Barret Schloerke \email{bigbear@iastate.edu}, Haesung Kim \email{hae0510@iastate.edu}, Di Cook \email{dicook@iastate.edu} and Heike Hofmann \email{hofmann@iastate.edu}}
\keyword{hplot}
\arguments{\item{data}{data set using}
\item{...}{}}
\examples{ggally_barDiag(movies, x = "mpaa")
ggally_barDiag(movies, x ="rating", binwidth = ".1")}
