% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{ggally_barDiag}
\alias{ggally_barDiag}
\title{Plots the Bar Plots by Using Diagonal}
\usage{
ggally_barDiag(data, mapping, ..., rescale = FALSE)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{...}{other arguments are sent to geom_bar}

\item{rescale}{boolean to decide whether or not to rescale the count output. Only applies to numeric data}
}
\description{
Plots the bar plots by using Diagonal.
}
\examples{
data(tips, package = "reshape")
ggally_barDiag(tips, mapping = ggplot2::aes(x = day))
ggally_barDiag(tips, mapping = ggplot2::aes(x = tip), binwidth = 0.25)
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}
