\name{ggally_blank}
\alias{ggally_blank}
\title{Blank...}
\usage{ggally_blank(...)}
\description{Blank
Drawing nothing}
\details{Makes a "blank" ggplot object that will only draw white space}
\author{Barret Schloerke \email{schloerke@gmail.com}}
\keyword{hplot}
\arguments{\item{...}{other arguments ignored}}
