\name{ggally_blank}
\alias{ggally_blank}
\title{Blank
Drawing nothing}
\usage{
  ggally_blank(...)
}
\arguments{
  \item{...}{other arguments ignored}
}
\description{
  Makes a "blank" ggplot object that will only draw white
  space
}
\author{
  Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}

