\name{ggally_blank}
\alias{ggally_blank}
\title{Blank...}

\description{
  Blank Drawing nothing
}

\details{
  Makes a "blank" ggplot object that will only draw white space
}
\author{Barret Schloerke \email{bigbear@iastate.edu}, Haesung Kim \email{hae0510@iastate.edu}, Di Cook \email{dicook@iastate.edu} and Heike Hofmann \email{hofmann@iastate.edu}}
\keyword{hplot}
