% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{ggally_blank}
\alias{ggally_blank}
\alias{ggally_blankDiag}
\alias{ggally_na}
\alias{ggally_naDiag}
\title{Blank}
\usage{
ggally_blank(...)

ggally_blankDiag(...)

ggally_na(data, mapping, size = 10, ...)

ggally_naDiag(...)
}
\arguments{
\item{...}{other arguments ignored}
}
\description{
Draws nothing.
}
\details{
Makes a "blank" ggplot object that will only draw white space
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}

