% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{ggally_blank}
\alias{ggally_blank}
\alias{ggally_blankDiag}
\title{Blank}
\usage{
ggally_blank(...)

ggally_blankDiag(...)
}
\arguments{
\item{...}{other arguments ignored}
}
\description{
Draws nothing.
}
\details{
Makes a "blank" ggplot object that will only draw white space
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}
