% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gg-plots.r
\name{ggally_blank}
\alias{ggally_blank}
\title{Blank}
\usage{
ggally_blank(...)
}
\arguments{
\item{...}{other arguments ignored}
}
\description{
Draws nothing.
}
\details{
Makes a "blank" ggplot object that will only draw white space
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}

