% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{ggally_count}
\alias{ggally_count}
\title{Plots the number of observations}
\usage{
ggally_count(data, mapping, ...)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{...}{other arguments passed to \code{\link[ggplot2]{geom_tile}(...)}}
}
\description{
Plot the number of observations by using rectangles
with proportional areas.
}
\examples{
data(tips, package = "reshape")
ggally_count(tips, mapping = ggplot2::aes(x = smoker, y = sex))

ggally_count(
  as.data.frame(Titanic),
  mapping = ggplot2::aes(x = Class, y = Survived, weight = Freq)
)
}
\author{
Joseph Larmarange \email{joseph@larmarange.net}
}
\keyword{hplot}
