% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{ggally_densityDiag}
\alias{ggally_densityDiag}
\title{Plots the Density Plots by Using Diagonal}
\usage{
ggally_densityDiag(data, mapping, ..., rescale = FALSE)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used.}

\item{...}{other arguments sent to stat_density}

\item{rescale}{boolean to decide whether or not to rescale the count output}
}
\description{
Plots the density plots by using Diagonal.
}
\examples{
 data(tips, package = "reshape")
 ggally_densityDiag(tips, mapping = ggplot2::aes(x = total_bill))
 ggally_densityDiag(tips, mapping = ggplot2::aes(x = total_bill, color = day))
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}
