% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gg-plots.r
\name{ggally_denstrip}
\alias{ggally_denstrip}
\title{Plots a tile plot with facets}
\usage{
ggally_denstrip(data, mapping, ...)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{...}{other arguments being sent to stat_bin}
}
\description{
Make Tile Plot as densely as possible.
}
\examples{
data(tips, package = "reshape")
 ggally_denstrip(tips, mapping = ggplot2::aes(x = total_bill, y = sex))
 ggally_denstrip(tips, mapping = ggplot2::aes_string(x = "total_bill", y = "sex"))
 ggally_denstrip(
   tips,
   mapping = ggplot2::aes_string(x = "sex", y = "tip", binwidth = "0.2")
 ) + ggplot2::scale_fill_gradient(low = "grey80", high = "black")
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}

