% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gg-plots.r
\name{ggally_dot}
\alias{ggally_dot}
\title{Plots the Box Plot with Dot}
\usage{
ggally_dot(data, mapping, ...)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{...}{other arguments being supplied to geom_jitter}
}
\description{
Add jittering with the box plot
}
\examples{
data(tips, package = "reshape")
 ggally_dot(tips, mapping = ggplot2::aes(x = total_bill, y = sex))
 ggally_dot(tips, mapping = ggplot2::aes_string(x = "total_bill", y = "sex"))
 ggally_dot(
   tips,
   mapping = ggplot2::aes_string(y = "total_bill", x = "sex", color = "sex")
 )
 ggally_dot(
   tips,
   mapping = ggplot2::aes_string(y = "total_bill", x = "sex", color = "sex", shape = "sex")
 ) + ggplot2::scale_shape(solid=FALSE)
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}

