% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{ggally_facetbar}
\alias{ggally_facetbar}
\title{Faceted bar plot}
\usage{
ggally_facetbar(data, mapping, ...)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{...}{other arguments are sent to geom_bar}
}
\description{
X variables are plotted using \code{geom_bar} and are faceted by the Y variable.
}
\examples{
 data(tips, package = "reshape")
 ggally_facetbar(tips, ggplot2::aes(x = sex, y = smoker, fill = time))
 ggally_facetbar(tips, ggplot2::aes(x = smoker, y = sex, fill = time))
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}
