% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{ggally_facetdensitystrip}
\alias{ggally_facetdensitystrip}
\title{Plots a density plot with facets or a tile plot with facets}
\usage{
ggally_facetdensitystrip(data, mapping, ..., den_strip = FALSE)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{...}{other arguments being sent to either geom_histogram or stat_density}

\item{den_strip}{boolean to decide whether or not to plot a density strip(TRUE) or a facet density(FALSE) plot.}
}
\description{
Make Tile Plot as densely as possible.
}
\examples{
example(ggally_facetdensity)
example(ggally_denstrip)
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}
