% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnostic.R
\name{ggally_nostic_cooksd}
\alias{ggally_nostic_cooksd}
\title{ggnostic function to display stats::cooks.distance}
\usage{
ggally_nostic_cooksd(data, mapping, ..., linePosition = 4/nrow(data),
  lineColor = brew_colors("purple"))
}
\arguments{
\item{data, mapping, ..., lineColor}{parameters supplied to \code{\link{ggally_nostic_line}}}

\item{linePosition}{1 is the general cutoff point for Cook's Distance}
}
\value{
ggplot2 plot object
}
\description{
A line is added at 1 to display the general cutoff point for Cook's Distance.
}
\details{
Reference: Cook, R. Dennis; Weisberg, Sanford (1982). Residuals and Influence in Regression. New York, NY: Chapman & Hall. ISBN 0-412-24280-X
}
\seealso{
\code{stats::\link[stats]{cooks.distance}}
}

