% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggduo_wrappers.R
\name{ggally_nostic_hat}
\alias{ggally_nostic_hat}
\title{ggnostic function to display stats::influence's hat information}
\usage{
ggally_nostic_hat(data, mapping, ..., linePosition = 2 *
  sum(data[[deparse(mapping$y)]])/nrow(data), lineColor = brew_colors$purple,
  lineSize = 0.5, lineAlpha = 1, lineType = 1,
  avgLinePosition = sum(data[[deparse(mapping$y)]])/nrow(data),
  avgLineColor = brew_colors$red, avgLineSize = lineSize,
  avgLineAlpha = lineAlpha, avgLineType = lineType)
}
\arguments{
\item{data, mapping, ...}{supplied directly to \code{\link{ggally_nostic_line}}}

\item{linePosition, lineColor, lineSize, lineAlpha, lineType}{parameters supplied to \code{ggplot2::\link[ggplot2]{geom_line}} for the cutoff line}

\item{avgLinePosition, avgLineColor, avgLineSize, avgLineAlpha, avgLineType}{parameters supplied to \code{ggplot2::\link[ggplot2]{geom_line}} for the average line}
}
\value{
ggplot2 plot object
}
\description{
As stated in \code{stats::\link[stats]{influence}} documentation:
}
\details{
hat: a vector containing the diagonal of the ‘hat’ matrix.

The diagonal elements of the 'hat' matrix describe the influence each response value has on the fitted value for that same observation.

A suggested "cutoff" line is added to the plot at a height of 2 * p / n and an expected line at a height of p / n.
If either \code{linePosition} or \code{avgLinePosition} is \code{NULL}, the respective line will not be drawn.
}
\seealso{
\code{stats::\link[stats]{influence}}
}

