% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggduo_wrappers.R
\name{ggally_nostic_resid}
\alias{ggally_nostic_resid}
\title{ggnostic function to display residuals}
\usage{
ggally_nostic_resid(data, mapping, ..., linePosition = 0,
  lineColor = brew_colors("red"), lineSize = 0.5, lineAlpha = 1,
  lineType = 1, lineConfColor = lineColor, lineConfSize = lineSize,
  lineConfAlpha = lineAlpha, lineConfType = 2, pVal = c(0.025, 0.975),
  sigma = attr(data, "broom_glance")$sigma)
}
\arguments{
\item{data, mapping, ...}{parameters supplied to \code{\link{ggally_nostic_line}}}

\item{linePosition, lineColor, lineSize, lineAlpha}{parameters supplied to \code{ggplot2::\link[ggplot2]{geom_line}}}

\item{lineConfColor, lineConfSize, lineConfAlpha, lineConfType}{parameters supplied to the confidence interval lines}

\item{pVal}{percentiles of a N(0, sigma) distribution to be drawn}

\item{sigma}{sigma value for the \code{pVal} percentiles}
}
\value{
ggplot2 plot object
}
\description{
If non-null \code{pVal} and \code{sigma} values are given, confidence interval lines will be added to the plot at the specified \code{pVal} percentiles of a N(0, sigma) distribution.
}
\seealso{
\code{stats::\link[stats]{residuals}}
}

