% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggduo_wrappers.R
\name{ggally_nostic_se_fit}
\alias{ggally_nostic_se_fit}
\title{ggnostic function to display stats::predict's standard errors}
\usage{
ggally_nostic_se_fit(data, mapping, ..., lineColor = brew_colors("green"),
  linePosition = 0)
}
\arguments{
\item{data, mapping, ..., lineColor}{parameters supplied to \code{\link{ggally_nostic_line}}}

\item{linePosition}{base comparison for a perfect fit}
}
\value{
ggplot2 plot object
}
\description{
As stated in \code{stats::\link[stats]{predict}} documentation:
}
\details{
If the logical ‘se.fit’ is ‘TRUE’, standard errors of the predictions are calculated.  If the numeric argument ‘scale’ is set (with optional ‘df’), it is used as the residual standard deviation in the computation of the standard errors, otherwise this is extracted from the model fit.

Since the se.fit is \code{TRUE} and scale is unset by default, the standard errors are extracted from the model fit.

A base line of 0 is added to give reference to a perfect fit.
}
\seealso{
\code{stats::\link[stats]{influence}}
}

