% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggduo_wrappers.R
\name{ggally_nostic_sigma}
\alias{ggally_nostic_sigma}
\title{ggnostic function to display stats::influence's sigma}
\usage{
ggally_nostic_sigma(data, mapping, ..., lineColor = brew_colors$red,
  linePosition = attr(data, "broom_glance")$sigma)
}
\arguments{
\item{data, mapping, ..., lineColor}{parameters supplied to \code{\link{ggally_nostic_line}}}

\item{sigma}{linePosition that is drawn in the background of the plot. Defaults to the overall model's sigma value.}
}
\value{
ggplot2 plot object
}
\description{
As stated in \code{stats::\link[stats]{influence}} documentation:
}
\details{
sigma: a vector whose i-th element contains the estimate of the residual standard deviation obtained when the i-th case is dropped from the regression.  (The approximations needed for GLMs can result in this being ‘NaN’.)

A line is added to display the overall model's sigma value. This gives a baseline for comparison
}
\seealso{
\code{stats::\link[stats]{influence}}
}

