% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gg-plots.r
\name{ggally_points}
\alias{ggally_points}
\title{Plots the Scatter Plot}
\usage{
ggally_points(data, mapping, ...)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{...}{other arguments are sent to geom_point}
}
\description{
Make a scatter plot with a given data set.
}
\examples{
data(mtcars)
ggally_points(mtcars, mapping = ggplot2::aes(x = disp, y = hp))
ggally_points(mtcars, mapping = ggplot2::aes_string(x = "disp", y = "hp"))
ggally_points(
  mtcars,
  mapping = ggplot2::aes_string(
    x     = "disp",
    y     = "hp",
    color = "as.factor(cyl)",
    size  = "gear"
  )
)
}
\author{
Barret Schloerke  \email{schloerke@gmail.com}
}
\keyword{hplot}

