% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg-plots.R
\name{ggally_ratio}
\alias{ggally_ratio}
\title{Plots a mosaic plots}
\usage{
ggally_ratio(data)
}
\arguments{
\item{data}{data set using}
}
\description{
Plots the mosaic plot by using fluctuation.
}
\details{
Must send only two discrete columns in the data set.
}
\examples{
data(tips, package = "reshape")
ggally_ratio(tips[,c("sex", "smoker")])
ggally_ratio(tips[,c("sex", "smoker")]) + ggplot2::coord_equal()
ggally_ratio(
  tips[,c("sex", "day")]
) + ggplot2::theme( aspect.ratio = 4/2)
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{hplot}

