% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_weighted_mean.R
\name{ggally_trends}
\alias{ggally_trends}
\title{Trends line plot}
\usage{
ggally_trends(data, mapping, ..., include_zero = FALSE)
}
\arguments{
\item{data}{data set using}

\item{mapping}{aesthetics being used}

\item{...}{other arguments passed to \code{\link[ggplot2]{geom_line}(...)}}

\item{include_zero}{Should 0 be included on the y-axis?}
}
\description{
Plot trends using line plots.
For continuous y variables, plot the evolution of the mean.
For binary y variables, plot the evolution of the proportion.
}
\examples{
data(tips, package = "reshape")
tips$day <- factor(tips$day, c("Thur", "Fri", "Sat", "Sun"))

# Numeric variable
ggally_trends(tips, mapping = aes(x = day, y = total_bill))
ggally_trends(tips, mapping = aes(x = day, y = total_bill, colour = time))

# Binary variable
ggally_trends(tips, mapping = aes(x = day, y = smoker))
ggally_trends(tips, mapping = aes(x = day, y = smoker, colour = sex))

# Discrete variable with 3 or more categories
ggally_trends(tips, mapping = aes(x = smoker, y = day))
ggally_trends(tips, mapping = aes(x = smoker, y = day, color = sex))

# Include zero on Y axis
ggally_trends(tips, mapping = aes(x = day, y = total_bill), include_zero = TRUE)
ggally_trends(tips, mapping = aes(x = day, y = smoker), include_zero = TRUE)

# Change line size
ggally_trends(tips, mapping = aes(x = day, y = smoker, colour = sex), size = 3)

# Define weights with the appropriate aesthetic
d <- as.data.frame(Titanic)
ggally_trends(
  d,
  mapping = aes(x = Class, y = Survived, weight = Freq, color = Sex),
  include_zero = TRUE
)


}
\author{
Joseph Larmarange \email{joseph@larmarange.net}
}
\keyword{hplot}
