% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs.R
\name{ggduo}
\alias{ggduo}
\title{ggduo - A ggplot2 generalized pairs plot for two columns sets of a data.frame}
\usage{
ggduo(data, mapping = NULL, columnsX = 1:ncol(data),
  columnsY = 1:ncol(data), types = list(continuous = "points", combo =
  "facethist", discrete = "ratio"), axisLabels = c("show", "none"),
  columnLabelsX = colnames(data[columnsX]),
  columnLabelsY = colnames(data[columnsY]), showStrips = NULL,
  legends = FALSE)
}
\arguments{
\item{data}{data set using.  Can have both numerical and categorical data.}

\item{mapping}{aesthetic mapping (besides \code{x} and \code{y}).  See \code{\link[ggplot2]{aes}()}.  If \code{mapping} is numeric, \code{columns} will be set to the \code{mapping} value and \code{mapping} will be set to \code{NULL}.}

\item{columnsX, columnsY}{which columns are used to make plots.  Defaults to all columns.}

\item{axisLabels}{either "show" to display axisLabels, "internal" for labels in the diagonal plots, or "none" for no axis labels}

\item{columnLabelsX, columnLabelsY}{label names to be displayed.  Defaults to names of columns being used.}

\item{showStrips}{boolean to determine if each plot's strips should be displayed. \code{NULL} will default to the top and right side plots only. \code{TRUE} or \code{FALSE} will turn all strips on or off respectively.}

\item{legends}{boolean to determine the printing of the legend in each plot. Not recommended.}

\item{title}{title for the graph}

\item{upper}{see Details}

\item{lower}{see Details}

\item{diag}{see Details}

\item{params}{deprecated.  Please see \code{\link{wrap_fn_with_param_arg}}}

\item{...}{other parameters being supplied to geom's aes, such as color}
}
\description{
Make a matrix of plots with a given data set with two different column sets
}

