% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmatrix_print_new.R
\name{ggmatrix_gtable}
\alias{ggmatrix_gtable}
\title{Print ggpair object}
\usage{
ggmatrix_gtable(x, ..., progress = interactive() && (x$ncol * x$nrow) > 15,
  progress_format = " plot: [:plot_i,:plot_j] [:bar]:percent est::eta\\n")
}
\arguments{
\item{x}{ggpair object to be plotted}

\item{...}{ignored}

\item{progress}{boolean to determine if a progress bar should be displayed. This defaults to interactive sessions only}

\item{progress_format}{string supplied directly to \code{progress::\link[progress]{progress_bar}(format = progress_format)}. Defaults to display the plot number, progress bar, percent complete, and estimated time to finish.}

\item{plot}{boolean to determine if the gtable should be drawn or returned}
}
\description{
Specialized method to print the ggpair object-
}
\examples{
 data(tips, package = "reshape")
 pm <- ggpairs(tips, c(1,3,2), mapping = ggplot2::aes_string(color = "sex"))
 ggprint(pm) # calls print(pm), which calls ggprint.ggmatrix(pm)
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}

