% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmatrix_gtable.R
\name{ggmatrix_gtable}
\alias{ggmatrix_gtable}
\title{Print ggmatrix object}
\usage{
ggmatrix_gtable(pm, ...)
}
\arguments{
\item{pm}{ggmatrix object to be plotted}

\item{...}{ignored}
}
\description{
Specialized method to print the ggmatrix object-
}
\examples{
data(tips, package = "reshape")
pm <- ggpairs(tips, c(1,3,2), mapping = ggplot2::aes_string(color = "sex"))
ggmatrix_gtable(pm)
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
