% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmatrix_print_new.R
\name{ggprint}
\alias{ggprint}
\title{Print ggpair object}
\usage{
ggprint(x, ..., progress = interactive(), progress_wait = 5, plot = TRUE)
}
\arguments{
\item{x}{ggpair object to be plotted}

\item{...}{ignored}

\item{progress}{boolean to determine if a progress bar should be displayed. This defaults to interactive sessions only}

\item{progress_wait}{how many seconds the progress bar will wait until appearing. Defaults to 5 seconds.}

\item{plot}{boolean to determine if the gtable should be drawn or returned}
}
\description{
Specialized method to print the ggpair object-
}
\examples{
 data(tips, package = "reshape")
 pm <- ggpairs(tips, c(1,3,2), mapping = ggplot2::aes_string(color = "sex"))
 ggprint(pm) # calls print(pm), which calls ggprint.ggmatrix(pm)
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}

