% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtable.R
\name{ggtable}
\alias{ggtable}
\title{Plot an outcome with several potential explanotory variables}
\usage{
ggtable(
  data,
  columnsX = 1:ncol(data),
  columnsY = 1:ncol(data),
  cells = c("observed", "prop", "row.prop", "col.prop", "expected", "residuals",
    "stdres"),
  fill = c("none", "stdres", "residuals"),
  mapping = NULL,
  ...
)
}
\arguments{
\item{data}{dataset to be used, can have both categorical and
numerical variables}

\item{columnsX, columnsY}{names or positions of which columns are used to make plots. Defaults to all columns.}

\item{cells}{Which statistic should be displayed in table cells?}

\item{fill}{Wich statistic should be used for filling table cells?}

\item{mapping}{additional aesthetic to be used, for example to indicate
weights (see examples)}

\item{...}{additional arguments passed to \code{\link{ggduo}} (see examples)}
}
\description{
\code{ggtable} is a variant of \code{\link{ggduo}} for plotting
an outcome variable with several potential explanatory variables.
}
\examples{
data(tips, package = "reshape")
ggtable(tips, "smoker", c("day", "time", "sex"))

# displaying row proportions
ggtable(tips, "smoker", c("day", "time", "sex"), cells = "row.prop")

# filling cells with standardized residuals
ggtable(tips, "smoker", c("day", "time", "sex"), fill = "stdres")

# if continuous variables are provided, just displaying some summary statistics
ggtable(tips, c("smoker", "total_bill"), c("day", "time", "sex", "tip"))
}
\author{
Joseph Larmarange \email{joseph@larmarange.net}
}
