% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnostic.R
\name{ggts}
\alias{ggts}
\title{Multiple Time Series}
\usage{
ggts(data, mapping, columnsX, columnsY, ...)
}
\arguments{
\item{data, mapping, columnsX, columnsY, ...}{supplied directly to \code{\link{ggduo}}}
}
\value{
ggmatrix object
}
\description{
GGally implementation of ts.plot. Wraps around the ggduo function and removes the column strips
}
\examples{
ggts(pigs, "time", c("gilts", "profit", "s_per_herdsz", "production", "herdsz"))
}

