% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggduo_wrappers.R
\name{ggts}
\alias{ggts}
\title{Multiple Time Series}
\usage{
ggts(data, mapping, columnsX, columnsY, ...)
}
\arguments{
\item{data, mapping, columnsX, columnsY, ...}{supplied directly to \code{\link{ggduo}}}
}
\value{
ggmatrix object
}
\description{
GGally implementation of ts.plot
}
\examples{
pigs <- read.table("http://lib.stat.cmu.edu/datasets/Andrews/T62.1")
pigs <- pigs[,-(1:3)]
colnames(pigs) <- c("year", "quarter", "gilts", "profit", "s_per_herdsz", "production", "herdsz")

pigs$time <- pigs$year + (pigs$quarter - 1) / 4

ggts(pigs, "time", c("gilts", "profit", "s_per_herdsz", "production", "herdsz"))
}

