% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnostic.R
\name{ggts}
\alias{ggts}
\title{Multiple Time Series}
\usage{
ggts(..., columnLabelsX = NULL, xlab = "time")
}
\arguments{
\item{...}{supplied directly to \code{\link{ggduo}}}

\item{columnLabelsX}{remove top strips for the X axis by default}

\item{xlab}{defaults to "time"}
}
\value{
ggmatrix object
}
\description{
GGally implementation of ts.plot. Wraps around the ggduo function and removes the column strips
}
\examples{
ggts(pigs, "time", c("gilts", "profit", "s_per_herdsz", "production", "herdsz"))
}
