% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggpairs.r
\name{mapping_color_fill}
\alias{mapping_color_fill}
\title{Aesthetic Mapping Color Fill}
\usage{
mapping_color_fill(current)
}
\arguments{
\item{current}{the current aesthetics}
}
\description{
Replace the fill with the color and make color NULL.
}
\keyword{internal}

