% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnostic.R
\name{model_response_variables}
\alias{model_beta_label}
\alias{model_beta_variables}
\alias{model_response_variables}
\title{Model term names}
\usage{
model_response_variables(model, data = broom::augment(model))

model_beta_variables(model, data = broom::augment(model))

model_beta_label(model, data = broom::augment(model))
}
\arguments{
\item{model}{model in question}

\item{data}{equivalent to \code{broom::augment(model)}}
}
\value{
character vector of names
}
\description{
Retrieve either the response variable names, the beta variable names, or beta variable names with signifigance stars.
}

