% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-combo.R
\name{plot_types}
\alias{plot_types}
\title{Plot Types}
\usage{
plot_types(data, columnsX, columnsY, allowDiag = TRUE)
}
\arguments{
\item{data}{data set to be used}
}
\description{
Retrieves the type of plot that should be used for all combinations
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{internal}
