% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/find-combo.r
\name{plot_types}
\alias{plot_types}
\title{Plot Types}
\usage{
plot_types(data)
}
\arguments{
\item{data}{data set to be used}
}
\description{
Retrieves the type of plot that should be used for all combinations
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{internal}

