% Generated by roxygen2 (4.0.1): do not edit by hand
\name{print.ggpairs}
\alias{print.ggpairs}
\title{Print ggpair object}
\usage{
\method{print}{ggpairs}(x, leftWidthProportion = 0.2,
  bottomHeightProportion = 0.1, spacingProportion = 0.03,
  showStrips = FALSE, ...)
}
\arguments{
\item{x}{ggpair object to be plotted}

\item{...}{not used}
}
\description{
Specialized method to print the ggpair object-
}
\examples{
data(tips, package = "reshape")
 pMat <- ggpairs(tips, 1:3, color = "sex")
 pMat
 print(pMat, leftWidthProportion = 3, spacingProportion = 1, bottomHeightProportion = 1, showStrips = TRUE)
}
\author{
Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{internal}

