\name{print.ggpairs}
\alias{print.ggpairs}
\title{Print ggpair object...}
\usage{\method{print}{ggpairs} (x, ...)}

\description{
  Print ggpair object Specialized method to print the ggapir object
}
\keyword{hplot}
\author{Barret Schloerke \email{bigbear@iastate.edu} and Haesung Kim \email{hae0510@iastate.edu}}
\arguments{
  \item{x}{ggpair object to be plotted}
  \item{...}{not used}
}
\examples{example(ggpair)}
