\name{print.ggpairs}
\alias{print.ggpairs}
\title{Print ggpair object
Specialized method to print the ggapir object}
\usage{
\method{print}{ggpairs}(x, ...)
}
\arguments{
  \item{x}{ggpair object to be plotted}

  \item{...}{not used}
}
\description{
  Print ggpair object Specialized method to print the
  ggapir object
}
\examples{
ggpairs(tips[, 1:3])
}
\author{
  Barret Schloerke \email{schloerke@gmail.com}
}
\keyword{internal}
