\name{putPlot}
\alias{putPlot}
\title{Put Plot...}
\usage{putPlot(plotMatrix, plotObj, rowFromTop, columnFromLeft)}

\description{
  Put Plot Function to place your own plot in the layout
}
\keyword{hplot}
\author{Barret Schloerke \email{bigbear@iastate.edu} and Haesung Kim \email{hae0510@iastate.edu}}
\arguments{
  \item{plotMatrix}{ggally object to be altered}
  \item{plotObj}{ggplot object to be placed}
  \item{rowFromTop}{row from the top}
  \item{columnFromLeft}{column from the left}
}
\examples{example(ggpairs)}
