% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gglyph.R
\name{rescale01}
\alias{rescale01}
\alias{range01}
\alias{max1}
\alias{mean0}
\alias{min0}
\alias{rescale01}
\alias{rescale11}
\title{Rescaling functions}
\usage{
range01(x)

max1(x)

mean0(x)

min0(x)

rescale01(x, xlim = NULL)

rescale11(x, xlim = NULL)
}
\arguments{
\item{x}{numeric vector}

\item{xlim}{value used in \code{range}}
}
\description{
Rescaling functions
}
