\name{skewness}
\alias{skewness}
\title{Calculate the sample skewness of a vector...}
\usage{skewness(x)}
\description{Calculate the sample skewness of a vector
while ignoring missing values.}
\author{Jason Crowley \email{crowley.jason.s@gmail.com}}
\value{sample skewness of \code{x}}
\arguments{\item{x}{numeric vector}}
