% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggparcoord.r
\name{skewness}
\alias{skewness}
\title{Sample skewness}
\usage{
skewness(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
sample skewness of \code{x}
}
\description{
Calculate the sample skewness of a vector
while ignoring missing values.
}
\author{
Jason Crowley \email{crowley.jason.s@gmail.com}
}

