% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{v1_ggmatrix_theme}
\alias{v1_ggmatrix_theme}
\title{Modify a \code{\link{ggmatrix}} object by adding an \pkg{ggplot2} object to all #'}
\usage{
v1_ggmatrix_theme()
}
\description{
Modify a \code{\link{ggmatrix}} object by adding an \pkg{ggplot2} object to all #'
}
\examples{

ggpairs(iris, 1:2) + v1_ggmatrix_theme()
# move the column names to the left and bottom
ggpairs(iris, 1:2, switch = "both") + v1_ggmatrix_theme()
}
