% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpairs_add.R
\name{v1_ggmatrix_theme}
\alias{v1_ggmatrix_theme}
\title{Modify a ggmatrix object by adding an ggplot2 object to all plots}
\usage{
v1_ggmatrix_theme()
}
\description{
Modify a ggmatrix object by adding an ggplot2 object to all plots
}
\examples{

ggpairs(iris, 1:2) + v1_ggmatrix_theme()
# move the column names to the left and bottom
ggpairs(iris, 1:2, switch = "both") + v1_ggmatrix_theme()
}
