% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggpairs.r
\name{vplayout}
\alias{vplayout}
\title{Viewport Layout Wrapper}
\usage{
vplayout(x, y)
}
\arguments{
\item{x}{row position}

\item{y}{coloumn position}
}
\description{
A wrapper function to set the viewport.
}
\author{
Hadley Wickham \email{h.wickham@gmail.com}
}
\keyword{internal}

